why is iter01 not resolving properly?
